requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.1
black==25.1.0
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.1
Django==5.2
django-location-field==2.7.3
django-rest-framework-nested==0.0.1
djangorestframework==3.16.0
fcm-django==2.2.1
firebase-admin==6.8.0
flower==2.0.1
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
httplib2==0.22.0
httptools==0.6.4
humanize==4.12.3
idna==3.10
kombu==5.5.3
msgpack==1.1.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
prometheus_client==0.21.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
swapper==1.4.0
tornado==6.5
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
